CRUD操作をsimul_search.goファイルに集約
